requirements.txt
Flask==2.3.3
Flask-WTF==1.1.1
Flask-Login==0.6.2
pymongo==4.5.0
python-dotenv==1.0.0
bcrypt==4.0.1
Pillow==10.0.0
Flask-Assets==2.1.0
WTForms-Components==0.10.5
gunicorn==21.2.0
dnspython==2.4.2
email-validator==2.0.0
python-slugify==8.0.1
Flask-Caching==2.0.2
python-magic==0.4.27
